requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.1.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.2
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
mysqlclient==2.0.1
oauthlib==3.1.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytube==10.1.0
pytube3==9.6.4
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
traitlets==5.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
